refactor(components): add type key to import
