+++ src/components/Alert/types.ts
@@ -1,4 +1,4 @@
-import { HTMLAttributes } from 'react';
+import type { HTMLAttributes } from 'react';
 
 export type AlertType = 'success' | 'error' | 'warning' | 'info';
