+++ src/components/Chip/types.ts
@@ -1,4 +1,4 @@
-import { HTMLAttributes } from 'react';
+import type { HTMLAttributes } from 'react';
 
 export interface ChipProps extends HTMLAttributes<HTMLDivElement> {
     label: string;
